changing to redux way and reducing the mess
