+++ apps/xprof_gui/priv/src/actions/index.js
@@ -3,6 +3,8 @@ export {
   getMonitoredFunctions,
   getFunctionsData,
   getFunctionsCalls,
+  setIDs,
+  setSize,
 } from './CollectingActions';
 export {
   stopMonitoringFunction,
