+++ apps/xprof_gui/priv/src/components/monitoring/Graph/Graph.jsx
@@ -1,57 +1,40 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import C3Chart from 'react-c3js';
-import 'c3/c3.css';
-import { AXIS, DATA, GRID, POINT, TRANSITION } from '../../../constants';
-import Grid from './Grid/Grid';
-
-const composeID = q => `${q.replace(/[^A-Za-z0-9_-]/g, '-')}`;
+import GridGraph from '../GridGraph';
+import LineGraph from '../LineGraph';
+import * as graph from '../../../constants/GraphTypes';
 
 const propTypes = {
   dps: PropTypes.arrayOf(PropTypes.object).isRequired,
   type: PropTypes.string.isRequired,
-  query: PropTypes.string.isRequired,
-  passId: PropTypes.number.isRequired,
+  associatedID: PropTypes.number.isRequired,
+  setSize: PropTypes.func.isRequired,
+  size: PropTypes.shape(PropTypes.any).isRequired,
 };
 
 const Graph = ({
   dps,
   type,
-  query,
-  passId,
+  associatedID,
+  setSize,
+  size,
 }) => {
-  const queryID = composeID(query);
-  const wrapperID = `graphWrapper-${queryID}`;
-  const wrapper = document.getElementById(wrapperID);
-  if (type === 'grid') {
-    return (
-      <div
-        id={wrapperID}
-        style={{
-          textAlign: 'center',
-          justifyContent: 'center',
-        }}
-      >
-        {wrapper && <Grid
-          data={{ ...DATA, json: dps }}
-          outerWidth={wrapper.clientWidth}
-          graphID={passId}
-        />}
-      </div>
-    );
+  switch (type) {
+    case graph.GRID:
+      return (
+        <GridGraph
+          dps={dps}
+          associatedID={associatedID}
+          setSize={setSize}
+          size={size}
+        />);
+    case graph.LINE:
+      return <LineGraph dps={dps} />;
+    default:
+      return null;
   }
-  return (
-    <div>
-      <C3Chart
-        data={{ ...DATA, json: dps }}
-        point={POINT}
-        grid={GRID}
-        axis={AXIS}
-        transition={TRANSITION}
-      />
-    </div>
-  );
 };
+
 Graph.propTypes = propTypes;
 
 export default Graph;
