+++ apps/xprof_gui/priv/src/components/monitoring/Grid/index.js
@@ -1,0 +1,1 @@
+export { default } from './Grid';
