+++ apps/xprof_gui/priv/src/components/monitoring/GridGraph/index.js
@@ -1,0 +1,1 @@
+export { default } from './GridGraph';
