+++ apps/xprof_gui/priv/src/components/monitoring/LineGraph/index.js
@@ -1,0 +1,1 @@
+export { default } from './LineGraph';
