+++ apps/xprof_gui/priv/src/constants/ActionTypes.js
@@ -21,6 +21,8 @@ export const STOP_MONITORING_FUNCTION = 'STOP_MONITORING_FUNCTION';
 export const STOP_MONITORING_FUNCTION_ERROR = 'STOP_MONITORING_FUNCTION_ERROR';
 export const EXPAND_GRAPH_PANEL = 'EXPAND_GRAPH_PANEL';
 export const SHRINK_GRAPH_PANEL = 'SHRINK_GRAPH_PANEL';
+export const ASSIGN_GRAPH_ID = 'ASSIGN_GRAPH_ID';
+export const CHANGE_SIZE = 'CHANGE_SIZE';
 
 // tracing
 export const UPDATE_CALLS = 'UPDATE_CALLS';
