+++ apps/xprof_gui/priv/src/constants/GraphTypes.js
@@ -1,0 +1,2 @@
+export const GRID = 'grid';
+export const LINE = 'line';
