+++ apps/xprof_gui/priv/src/constants/index.js
@@ -3,3 +3,4 @@ export * from './UIConstants';
 export * from './TracingConstants';
 export * from './CollectingConstants';
 export * from './GraphConstants';
+export * from './GraphTypes';
