+++ apps/xprof_gui/priv/src/utils/index.js
@@ -2,3 +2,5 @@
 export { default } from './registerServiceWorker';
 export * from './ActionsUtils';
 export * from './CommonUtils';
+export * from './gridComposingUtils';
+export * from './gridUtils';
